Scons build for enuminst, 64-bit w/Ragel only.
Whitespace fix.

Add scons build for enuminst, 64-bit w/Ragel only.

Tweak enuminst for ragel build.

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10545102

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8887 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 2a79f2a4b64d27fef92d2f2e09ea7b9a245d30be

+++ src/trusted/validator_x86/ncenuminsts.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011 The Native Client Authors. All rights reserved.
+ * Copyright (c) 2012 The Native Client Authors. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
  */
@@ -88,7 +88,7 @@ Bool NaClInstValidates(uint8_t* mbase,
  *    vbase - The virtual address associated with the instruction.
  */
 Bool NaClSegmentValidates(uint8_t* mbase,
-                          uint8_t size,
+                          size_t size,
                           NaClPcAddress vbase);
 
 #endif  /* NATIVE_CLIENT_SRC_TRUSTED_VALIDATOR_X86_NCEMNUMINSTS_H_ */
